gravitee-policy-quota/src/main/java/io/gravitee/policy/quota/configuration/QuotaPolicyConfiguration.java
/**
 * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *         http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package io.gravitee.policy.quota.configuration;

import io.gravitee.policy.api.PolicyConfiguration;

/**
 * @author David BRASSELY (david.brassely at graviteesource.com)
 * @author GraviteeSource Team
 */
public class QuotaPolicyConfiguration implements PolicyConfiguration {

    private boolean async;

    private boolean addHeaders = true;

    private String key;

    private QuotaConfiguration quota;

    public QuotaConfiguration getQuota() {
        return quota;
    }

    public void setQuota(QuotaConfiguration quota) {
        this.quota = quota;
    }

    public boolean isAsync() {
        return async;
    }

    public void setAsync(boolean async) {
        this.async = async;
    }

    public boolean isAddHeaders() {
        return addHeaders;
    }

    public void setAddHeaders(boolean addHeaders) {
        this.addHeaders = addHeaders;
    }

    public String getKey() {
        return key;
    }

    public void setKey(String key) {
        this.key = key;
    }
}
